refactor: Move mask nodes to individual files
Implements ML-597
